+++ tool-appgallery/src/main/kotlin/com/infinum/sentinel/ui/tools/AppGalleryTool.kt
@@ -9,14 +9,17 @@ import com.infinum.sentinel.R
 import com.infinum.sentinel.Sentinel
 
 /**
- * Specific wrapper tool around Google Play.
+ * Specific wrapper tool around Huawei AppGallery.
+ *
+ * @param appId is your Huawei application ID from Huawei developer console.
  *
  * Tool Activity will launch with FLAG_ACTIVITY_RESET_TASK_IF_NEEDED, FLAG_ACTIVITY_CLEAR_TOP
  * and FLAG_ACTIVITY_NEW_TASK flags.
- * If no appropriate application is found, this tool will open a website on play.google.com.
+ * If no appropriate application is found, this tool will open a website on appgallery.cloud.huawei.com.
  */
 @SuppressLint("QueryPermissionsNeeded")
-public data class GooglePlayTool(
+public data class AppGalleryTool(
+    private val appId: String,
     private val listener: View.OnClickListener = View.OnClickListener { view ->
         val intent = Intent(
             Intent.ACTION_VIEW,
@@ -51,10 +54,9 @@ public data class GooglePlayTool(
                 Uri.Builder()
                     .scheme(SCHEME_HTTPS)
                     .authority(AUTHORITY)
-                    .appendPath(PATH_STORE)
-                    .appendPath(PATH_APPS)
-                    .appendPath(PATH_DETAILS)
-                    .appendQueryParameter(QUERY_ID, view.context.packageName)
+                    .appendPath(PATH_MARKETSHARE)
+                    .appendPath(PATH_APP)
+                    .appendPath("${PATH_C}$appId")
                     .build()
             )
         )
@@ -62,15 +64,16 @@ public data class GooglePlayTool(
 ) : Sentinel.DistributionTool {
 
     internal companion object {
-        private const val SCHEME_MARKET = "market"
+        private const val SCHEME_MARKET = "appmarket"
         private const val PATH_DETAILS = "details"
         private const val QUERY_ID = "id"
-        private const val VENDING_PACKAGE_NAME = "com.android.vending"
+        private const val VENDING_PACKAGE_NAME = "com.huawei.appmarket"
 
         private const val SCHEME_HTTPS = "https"
-        private const val AUTHORITY = "play.google.com"
-        private const val PATH_STORE = "store"
-        private const val PATH_APPS = "apps"
+        private const val AUTHORITY = "appgallery.cloud.huawei.com"
+        private const val PATH_MARKETSHARE = "marketshare"
+        private const val PATH_APP = "app"
+        private const val PATH_C = "C"
     }
 
     /**
@@ -78,7 +81,7 @@ public data class GooglePlayTool(
      *
      * @return a String resource that will be used to generate a name for a Button in Tools UI
      */
-    override fun name(): Int = R.string.sentinel_google_play
+    override fun name(): Int = R.string.sentinel_app_gallery
 
     /**
      * A callback to be invoked when this view is clicked.